pyproject.toml
[project]
name = "ept"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.11",
    "passlib>=1.7.4",
    "python-jose>=3.4.0",
    "ruff>=0.9.10",
    "sqlmodel>=0.0.24",
]
